[driver] invalidate data cache on data access reopens
test cases 208, 219, 221 of fileio were failing if rdbss caching was
turned on. the reason was, on reopen of a file, rdbss had cached,
unflushed data yet so reopen had incorrect view of the file.

need to invalidate the cache even if the open didn't return a
different change attribute than the one stored in the fcb, unfortunately.