dont set size of open for named attrs file
opens for create unchecked shouldn't have a non-zero size. if it does
it will either be ignored or flagged as an error.

casey notes that every open of named attributes is a create open.
even though we know the size of the file we are about to write,
we need to truncate it first, as setting the size will be ignored
by the server, leading to keeping the old size.